Separate data from checkpoints.cpp
